refactor: Refactor: split orbit-knowledge graph/sqlite_index.rs into fo… [ORB-00121]
Refactor: split orbit-knowledge graph/sqlite_index.rs into focused submodules

Planned-By: grok
